#13 - 기능 메서드 모듈화
- 포켓몬 관련 데이터를 불러오는 부분을 다른 파일로 분리해서 관리
